feat: Merge actors and directors page into movie detail
