--- movies/serializers/director.py
@@ -1,9 +1,0 @@
-from rest_framework import serializers
-from ..models import Director
-
-
-class DirectorSerializer(serializers.ModelSerializer):
-
-    class Meta:
-        model = Director
-        fields = '__all__'
