+++ movies/serializers/movie.py
@@ -21,9 +21,25 @@ class MovieSerializer(serializers.ModelSerializer):
             model = Director
             fields = '__all__'
 
+    class CastingSerializer(serializers.ModelSerializer):
+
+        class ActorSerializer(serializers.ModelSerializer):
+            class Meta:
+                model = Actor
+                fields = '__all__'
+
+        actor = ActorSerializer(read_only=True)
+
+        class Meta:
+            model = Casting
+            fields = '__all__'
+
+    castings = CastingSerializer(many=True, read_only=True)
     genre_ids = GenreSerializer(many=True, read_only=True)
     directors = DirectorSerializer(many=True, read_only=True)
 
     class Meta:
         model = Movie
-        fields = '__all__'
+        fields = ('id', 'title', 'adult', 'release_date', 'popularity', 'vote_count', 'vote_average',
+                  'overview', 'original_language', 'original_title', 'genre_ids', 'directors',
+                  'backdrop_path', 'poster_path', 'videos', 'recommendations', 'castings')
