+++ movies/urls.py
@@ -5,9 +5,6 @@ urlpatterns = [
     path('movies/', views.movie_list),
     path('movies/<int:movie_pk>/', views.movie_detail),
 
-    path('movies/<int:movie_pk>/actors', views.movie_detail_actors),
-    path('movies/<int:movie_pk>/directors', views.movie_detail_directors),
-
     path('movies/<int:movie_pk>/reviews/', views.review_create_or_list),
     path('movies/<int:movie_pk>/reviews/<int:review_pk>/', views.review_detail_or_update_delete),
