+++ movies/views.py
@@ -8,9 +8,7 @@ from rest_framework.permissions import AllowAny
 from rest_framework.response import Response
 from rest_framework.renderers import JSONRenderer
 
-from .models import Movie, Review, Comment
-from .serializers.actor import ActorSerializer
-from .serializers.director import DirectorSerializer
+from .models import Movie, Actor, Review, Comment
 from .serializers.movie import MovieSerializer, MovieListSerializer
 from .serializers.review import ReviewSerializer, CommentSerializer
 
@@ -31,24 +29,6 @@ def movie_list(request):
 def movie_detail(request, movie_pk):
     movie = get_object_or_404(Movie, pk=movie_pk)
     serializer = MovieSerializer(movie)
-    return Response(serializer.data)
-
-
-@api_view(['GET'])
-@permission_classes([AllowAny])
-def movie_detail_actors(request, movie_pk):
-    movie = get_object_or_404(Movie, pk=movie_pk)
-    actors = movie.actors.all()
-    serializer = ActorSerializer(actors, many=True)
-    return Response(serializer.data)
-
-
-@api_view(['GET'])
-@permission_classes([AllowAny])
-def movie_detail_directors(request, movie_pk):
-    movie = get_object_or_404(Movie, pk=movie_pk)
-    directors = movie.directors.all()
-    serializer = DirectorSerializer(directors, many=True)
     return Response(serializer.data)
